// server/src/boost/dto/payment.dto.ts
export class IPNPaymentDto {
  ipn_version: any;
  ipn_type: any;
  ipn_mode: any;
  ipn_id: any;
  merchant: any;
  deposit_id: any;
  txn_id: any;
  address: any;
  status: any;
  status_text: any;
  currency: any;
  confirms: any;
  amount: any;
  amounti: any;
  fee: any;
  feei: any;
  fiat_coin: any;
  fiat_amount: any;
  fiat_amounti: any;
  fiat_fee: any;
  fiat_feei: any;
}